code review changes and date sort fixed
